Rewrote some parts of the resource management, to make it possible to transparently read real files instead of datafiles

+++ src/OptionMenu.cc
@@ -1,4 +1,4 @@
-//  $Id: OptionMenu.cc,v 1.17 2000/06/23 17:05:43 grumbel Exp $
+//  $Id: OptionMenu.cc,v 1.18 2000/06/25 20:22:18 grumbel Exp $
 //
 //  Pingus - A free Lemmings clone
 //  Copyright (C) 1999 Ingo Ruhnke <grumbel@gmx.de>
@@ -263,8 +263,8 @@ OptionMenu::draw()
   }
 
   back->put_screen(0, 0);
-  if (CL_Mouse::get_x() < back->get_width()
-      && CL_Mouse::get_y() < back->get_height())
+  if (CL_Mouse::get_x() < (int)back->get_width()
+      && CL_Mouse::get_y() < (int)back->get_height())
     CL_Display::fill_rect(0, 0, back->get_width(), back->get_height(),
 			  1.0, 1.0, 1.0, 0.3);
   Display::flip_display();
@@ -322,8 +322,8 @@ OptionMenu::display()
 
       if (CL_Mouse::left_pressed()) 
 	{
-	  if (CL_Mouse::get_x() < back->get_width() &&
-	      CL_Mouse::get_y() < back->get_height())
+	  if (CL_Mouse::get_x() < (int)back->get_width() &&
+	      CL_Mouse::get_y() < (int)back->get_height())
 	    {
 	      quit = true;	      
 	    }
