+++ src/ThemeSelector.cc
@@ -1,4 +1,4 @@
-//  $Id: ThemeSelector.cc,v 1.23 2000/06/23 18:39:56 grumbel Exp $
+//  $Id: ThemeSelector.cc,v 1.24 2000/06/25 20:22:18 grumbel Exp $
 //
 //  Pingus - A free Lemmings clone
 //  Copyright (C) 1999 Ingo Ruhnke <grumbel@gmx.de>
@@ -270,8 +270,8 @@ ThemeSelector::draw()
   right_arrow->put_screen(CL_Display::get_width() - right_arrow->get_width(),
 			  (CL_Display::get_height() - right_arrow->get_height()) / 2);
   back->put_screen(0, 0);
-  if (CL_Mouse::get_x() < back->get_width()
-      && CL_Mouse::get_y() < back->get_height())
+  if (CL_Mouse::get_x() < (int)back->get_width()
+      && CL_Mouse::get_y() < (int)back->get_height())
     CL_Display::fill_rect(0, 0, back->get_width(), back->get_height(),
 			  1.0, 1.0, 1.0, 0.3);
